# BackendAPI/interfaces/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Robot Framework Test Manager API",
    "description": "RESTful API for managing Robot Framework test scripts, test cases, execution, queue, run history, and logs. Secured with JWT authentication and role-based access control.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/tests": {
      "post": {
        "tags": [
          "Test Scripts"
        ],
        "summary": "Create a new test script",
        "description": "Create a new test script.\n\n- **name**: Name of the test script (required)\n- **description**: Optional description\n- **metadata**: Optional metadata as JSON object\n\nReturns the created test script with ID and timestamps.",
        "operationId": "create_test_api_v1_tests_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestScriptCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Test Scripts"
        ],
        "summary": "List all test scripts",
        "description": "List all test scripts with pagination.\n\n- **page**: Page number (default: 1)\n- **page_size**: Items per page (default: 20, max: 100)\n\nReturns paginated list of test scripts.",
        "operationId": "list_tests_api_v1_tests_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tests/{test_id}": {
      "get": {
        "tags": [
          "Test Scripts"
        ],
        "summary": "Retrieve details of a specific test script",
        "description": "Get a specific test script by ID.\n\n- **test_id**: UUID of the test script\n\nReturns test script details including metadata and timestamps.",
        "operationId": "get_test_api_v1_tests__test_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "test_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Test Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestScriptResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Test Scripts"
        ],
        "summary": "Update a test script",
        "description": "Update an existing test script.\n\n- **test_id**: UUID of the test script\n- **name**: Optional new name\n- **description**: Optional new description\n- **metadata**: Optional new metadata\n\nReturns the updated test script.",
        "operationId": "update_test_api_v1_tests__test_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "test_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Test Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestScriptUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Test Scripts"
        ],
        "summary": "Delete a test script",
        "description": "Delete a test script.\n\n- **test_id**: UUID of the test script\n\nDeletes the test script and all associated test cases.\nReturns 204 No Content on success.",
        "operationId": "delete_test_api_v1_tests__test_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "test_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Test Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/queue": {
      "get": {
        "tags": [
          "Queue"
        ],
        "summary": "List currently queued test cases",
        "description": "Retrieve the current execution queue.\n\nReturns a list of queued test cases ordered by priority and queue time.\nEach item includes case ID, status, priority, and queued timestamp.",
        "operationId": "get_queue_api_v1_queue_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QueueItemResponse"
                  },
                  "type": "array",
                  "title": "Response Get Queue Api V1 Queue Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Queue"
        ],
        "summary": "Add test case(s) to queue",
        "description": "Add one or more test cases to the execution queue.\n\n- **case_ids**: List of test case UUIDs to add to queue\n\nReturns list of created queue items.\nSkips test cases that are already in the queue.",
        "operationId": "add_to_queue_api_v1_queue_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueueAddRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/queue/{case_id}": {
      "delete": {
        "tags": [
          "Queue"
        ],
        "summary": "Remove a test case from queue",
        "description": "Remove a test case from the execution queue.\n\n- **case_id**: UUID of the test case to remove\n\nReturns 204 No Content on success.\nOnly removes pending queue items.",
        "operationId": "remove_from_queue_api_v1_queue__case_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginationResponse": {
        "properties": {
          "items": {
            "items": {},
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "PaginationResponse",
        "description": "Schema for paginated responses"
      },
      "QueueAddRequest": {
        "properties": {
          "case_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Case Ids"
          }
        },
        "type": "object",
        "required": [
          "case_ids"
        ],
        "title": "QueueAddRequest",
        "description": "Schema for adding items to queue"
      },
      "QueueItemResponse": {
        "properties": {
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "queued_at": {
            "type": "string",
            "format": "date-time",
            "title": "Queued At"
          },
          "priority": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "case_id",
          "status",
          "queued_at"
        ],
        "title": "QueueItemResponse",
        "description": "Schema for queue item response"
      },
...
      "StandardResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "StandardResponse",
        "description": "Standard API response schema"
      },
...
      "TestScriptCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "meta_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta Data",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "TestScriptCreate",
        "description": "Schema for creating a test script"
      },
      "TestScriptResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "meta_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta Data",
            "default": {}
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "TestScriptResponse",
        "description": "Schema for test script response"
      },
      "TestScriptUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "meta_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta Data"
          }
        },
        "type": "object",
        "title": "TestScriptUpdate",
        "description": "Schema for updating a test script"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}